Refactor inventory.hasGood to use GoodName instead of Good.
Currently hasGood extracts the GoodName out of Good to check if the inventory has a good with the same GoodName. This creates additional baggage. If one wanted to check if a particular GoodName was in the inventory, one would have to wrap that GoodName in a dummy Good to call the hasGood method.